:bug: Add configuration file opening and add exit on service creation failure
The commit introduces opening and dumping configuration file. This is done in order to ensure that the configuration file is accessed properly and expected initial state is dumped into it. Furthermore, os.Exit(2) has been added on the service creation failure for a robust error handling. The comment on the SaveFileOnSet function has also been removed and we have simplified the error logging mechanism.